Extract typed entity and util date helper to module
