+++ bankapp/build.gradle
@@ -2,4 +2,5 @@ apply plugin: 'war'
 
 dependencies {
     compile project (':pubsub')
+    compile project (':entityhelper')
 }
