+++ bankapp/src/main/kotlin/com/clouway/bankapp/adapter/gae/datastore/DatastoreTransactionRepository.kt
@@ -2,6 +2,7 @@ package com.clouway.bankapp.adapter.gae.datastore
 
 import com.clouway.bankapp.core.*
 import com.clouway.bankapp.core.Transaction
+import com.clouway.entityhelper.TypedEntity
 import com.google.appengine.api.datastore.*
 import com.google.appengine.api.datastore.FetchOptions.Builder.withLimit
 import java.time.LocalDateTime
