+++ settings.gradle
@@ -2,4 +2,5 @@ rootProject.name = 'webbank'
 include 'bankapp'
 include 'mailservice'
 include 'pubsub'
+include 'entityhelper'
